Small fixes and cleanup
